Move modules from cutty.api to cutty.core
